Split benchmark suite for BLAS operators
