+++ benchmark/test_mv.py
@@ -1,0 +1,33 @@
+from typing import Generator
+
+import pytest
+import torch
+
+from . import performance_utils as base
+from .attri_util import FLOAT_DTYPES
+
+
+class MvBenchmark(base.GenericBenchmark2DOnly):
+    def set_more_shapes(self):
+        return []
+
+    def get_input_iter(self, dtype) -> Generator:
+        for m, n in self.shapes:
+            yield from self.input_fn(m, n, dtype, self.device)
+
+
+def _input_fn(m, n, cur_dtype, device):
+    inp1 = torch.randn([m, n], dtype=cur_dtype, device=device)
+    inp2 = torch.randn([n], dtype=cur_dtype, device=device)
+    yield inp1, inp2
+
+
+@pytest.mark.mv
+def test_mv():
+    bench = MvBenchmark(
+        op_name="mv",
+        input_fn=_input_fn,
+        torch_op=torch.Tensor.mv,
+        dtypes=FLOAT_DTYPES,
+    )
+    bench.run()
